requirements.txt
requests~=2.28.1
beautifulsoup4~=4.11.1
click~=8.1.3
python-dateutil~=2.8.2
tqdm~=4.64.1
urllib3~=1.26.13
aenum~=3.1.11